Revert "Autenticar sesión (client-side) mediante Firebase Auth"
