+++ src/nutrimetas/app.json
@@ -33,9 +33,7 @@
     },
     "plugins": [
       "expo-router",
-      "@react-native-firebase/app",
-      "@react-native-firebase/auth",
-      "expo-font"
+      "@react-native-firebase/app"
     ],
     "experiments": {
       "typedRoutes": true
