--- src/nutrimetas/app/(app)/(root)/addComment.tsx
@@ -1,6 +1,0 @@
-const addComment = () => {
-
-
-}
-
-export default addComment
