--- src/nutrimetas/app/(app)/(root)/index.tsx
@@ -1,5 +1,0 @@
-import { Text } from "react-native";
-
-export default function app() {
-    return <> <Text> hey </Text> </>;
-}
