+++ src/nutrimetas/app/(tabs)/_layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import FontAwesome from '@expo/vector-icons/FontAwesome';
-import { Tabs } from 'expo-router';
-import {Text, StyleSheet } from 'react-native';
+import { Link, Tabs } from 'expo-router';
+import { Pressable } from 'react-native';
 
 import Colors from '@/constants/Colors';
 import { useColorScheme } from '@/components/useColorScheme';
@@ -15,14 +15,6 @@ function TabBarIcon(props: {
   return <FontAwesome size={28} style={{ marginBottom: -3 }} {...props} />;
 }
 
-function ExpedientesScreen() {
-  return (
-    <Text style={styles.subtitle}>
-      NUTRI<Text style={{ color: Colors.lightblue }}>METAS</Text>
-    </Text>
-  );
-}
-
 export default function TabLayout() {
   const colorScheme = useColorScheme();
 
@@ -30,33 +22,24 @@ export default function TabLayout() {
     <Tabs
       screenOptions={{
         tabBarActiveTintColor: Colors[colorScheme ?? 'light'].tint,
+        // Disable the static render of the header on web
+        // to prevent a hydration error in React Navigation v6.
         headerShown: useClientOnlyValue(false, true),
-        headerTitleAlign: 'center',
       }}>
       <Tabs.Screen
         name="expedientes"
-        options={{ 
-          headerTitle: () => <ExpedientesScreen />, 
+        options={{
+          title: 'Expedientes',
           tabBarIcon: ({ color }) => <TabBarIcon name="address-book" color={color} />,
         }}
       />
       <Tabs.Screen
         name="goals"
         options={{
-          title: 'Mis Plantillas',
+          title: 'Goals',
           tabBarIcon: ({ color }) => <TabBarIcon name="flag" color={color} />,
         }}
       />
     </Tabs>
   );
 }
-
-
-const styles = StyleSheet.create({
-  subtitle: {
-    fontSize: 30,
-    fontWeight: 'bold',
-    color: Colors.green,
-    top: -20,
-  },
-});
