+++ src/nutrimetas/app/(tabs)/expedientes.tsx
@@ -1,4 +1,4 @@
-import { StyleSheet, View, Text } from 'react-native';
+import { ScrollView, StyleSheet, TouchableOpacity, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { Link } from 'expo-router';
 import Colors from '@/constants/Colors';
@@ -8,7 +8,6 @@ export default function Expedientes() {
   return (
     <SafeAreaView style={styles.container}>
       <View style={styles.listContainer}>
-      <Text style={styles.title}>Expedientes</Text>
         <PatientList />
       </View>
       <Link push href="/addPatient" style={styles.addButton}>
@@ -21,31 +20,24 @@ export default function Expedientes() {
 const styles = StyleSheet.create({
   container: {
     height: '100%',
-    backgroundColor: Colors.lightblue,
+    backgroundColor: "white"
   },
   listContainer: {
-    backgroundColor: Colors.white,
+    backgroundColor: "#FFFFFF",
     paddingHorizontal: 20,
-    top: -73,
-    height: '103%',
+    top: -50
   },
   addButton: {
     backgroundColor: Colors.lightblue,
     fontWeight: "bold",
     width: '100%',
-    height: 55,
+    height: 50,
     position: "absolute",
     bottom: 0,
     justifyContent: "center",
     textAlign: 'center',
-    color: Colors.white,
+    color: 'white',
     fontSize: 16,
     lineHeight: 50
-  },
-  title: {
-    marginVertical: 20,
-    fontSize: 36,
-    fontWeight: 'bold',
-    backgroundColor: Colors.white,
   },
 });
