+++ src/nutrimetas/app/(tabs)/goals.tsx
@@ -1,18 +1,18 @@
-import { StyleSheet, View } from 'react-native';
+import { StyleSheet, TouchableOpacity } from 'react-native';
+import { Text, View } from '@/components/Themed';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { Link } from 'expo-router';
 import Colors from '@/constants/Colors';
-import TemplatedGoals from '../templatedGoals';
+import InfoGoals from '../assingGoal';
+import FlashMessage from "react-native-flash-message";
 
 export default function Goals() {
   return (
     <SafeAreaView style={styles.container}>
-      <View style={styles.templateGoalList}>
-        <TemplatedGoals />
-      </View>
       <Link push href="/assingGoal" style={styles.button}>
             Crear Meta
       </Link>
+      <FlashMessage position="top" />
     </SafeAreaView>
   );
 }
@@ -22,7 +22,6 @@ const styles = StyleSheet.create({
     flex: 1,
     alignItems: 'center',
     justifyContent: 'center',
-    backgroundColor: Colors.white,
   },
   button: {
     backgroundColor: Colors.lightblue,
@@ -36,8 +35,5 @@ const styles = StyleSheet.create({
     color: 'white',
     fontSize: 16,
     lineHeight: 50
-  },
-  templateGoalList: {
-    top: -50,
   }
 });
