+++ src/nutrimetas/components/LoginForm.tsx
@@ -1,20 +1,16 @@
+"use strict"
+
 // Dependencies
 // React Hooks & References
-import { useRef, ElementRef } from "react";
+import { useRef } from "react";
 
 // Core React Native UI
-import { StyleSheet, Pressable } from "react-native";
+import { View, Text, TextInput, StyleSheet, Pressable, 
+    ImageSourcePropType } from "react-native";
 
 // Expo UI
+import { useAssets } from 'expo-asset'; 
 import { Image } from "expo-image";
-
-// Color palettes
-import Colors from "@/constants/Colors";
-import { View, Text, TextInput } from "@/components/Themed";
-
-// Image assets
-import MailIcon from '@/assets/images/mail.svg';
-import LockIcon from '@/assets/images/lock.svg';
 
 // Form structure and hooks
 import { useForm, SubmitHandler, Controller, Form } from "react-hook-form";
@@ -65,9 +61,15 @@ export default function LoginForm(
 
     // Register the login form's field refs
     const {emailRef, passwordRef} = {
-        emailRef: useRef<ElementRef<typeof TextInput>>(null),
-        passwordRef: useRef<ElementRef<typeof TextInput>>(null),
-    } as const; // 
+        emailRef: useRef<TextInput>(null),
+        passwordRef: useRef<TextInput>(null),
+    } as const;
+
+    // Register the icon loading hook
+    const [icon, error] = useAssets([
+        require('@/assets/images/mail.svg'), 
+        require('@/assets/images/lock.svg')
+    ]);
 
     // Render login form
     return (
@@ -79,10 +81,8 @@ export default function LoginForm(
                 {/* Input */}
                 <View style={FormStyles.FieldInputView}>
                     <Image 
-                        source={MailIcon}
-                        onError={(error) => {
-                            console.error("Error loading image:", error);
-                        }}
+                        source={icon? icon[0] as ImageSourcePropType : undefined}
+                        onError={() => {console.error("Error loading image:", error);}}
 
                         style={FormStyles.InputTextIcon}
                         contentFit="contain"
@@ -129,10 +129,9 @@ export default function LoginForm(
                 {/* Input */}
                 <View style={FormStyles.FieldInputView}>
                     <Image 
-                        source={LockIcon}
-                        onError={(error) => {
-                            console.error("Error loading image:", error);
-                        }}
+                        source={icon? icon[1] as ImageSourcePropType : undefined}
+                        onError={() => {
+                            console.error("Error loading image:", error);}}
 
                         style={FormStyles.InputTextIcon}
                         contentFit="contain"
@@ -219,13 +218,13 @@ const FormStyles = StyleSheet.create({
         justifyContent: "flex-start",
 
         borderBottomWidth: 1,
-        borderBottomColor: Colors.gray,
+        borderBottomColor: '#A6A6A6',
         // backgroundColor: "green"
     },
     InputTextIcon: {
         width: 25,
         height: 25,
-        color: Colors.gray,
+        color: '#A6A6A6',
     },
     FieldInputText: {
         fontWeight: "normal",
@@ -238,7 +237,7 @@ const FormStyles = StyleSheet.create({
         fontWeight: "bold",
         fontFamily: "sans-serif-light",
         fontStyle: "italic",
-        color: Colors.red,
+        color: "red",
 
         textAlign: "left",
     },
@@ -251,15 +250,15 @@ const FormStyles = StyleSheet.create({
         alignItems: 'center',
         justifyContent: 'center',
         
-        backgroundColor: Colors.lightblue,
+        backgroundColor: '#00C0F3',
     },
     LoginButtonText: {
         fontWeight: "bold",
         fontFamily: "sans-serif-light",
         fontStyle: "normal",
-        color: Colors.white,
+        color: "white",
 
         textAlign: "justify",
-        borderBottomColor: Colors.gray,
+        borderBottomColor: '#A6A6A6',
     },
 });
