+++ src/nutrimetas/constants/Colors.ts
@@ -4,7 +4,6 @@ const tintColorDark = '#fff';
 export default {
   light: {
     text: '#000',
-    placeholderText: '#a9a9ac',
     background: '#fff',
     tint: tintColorLight,
     tabIconDefault: '#ccc',
@@ -12,7 +11,6 @@ export default {
   },
   dark: {
     text: '#fff',
-    placeholderText: '#a9a9ac',
     background: '#000',
     tint: tintColorDark,
     tabIconDefault: '#ccc',
@@ -21,12 +19,5 @@ export default {
   yellow: "#ffe06a",
   lightblue: "#00c0f3",
   blue: "#005da4",
-  green: "#6dc067",
-  greenUCR: "#8dc63f",
-  white: "#FFFFFF",
-  gray: "#808080",
-  lightGray: '#ccc',
-  transparent: '#ffffff00',
-  red: '#FF0000',
-  black: '#000000'
+  green: "#6dc067"
 };
