+++ src/nutrimetas/package.json
@@ -16,14 +16,11 @@
     "@expo/vector-icons": "^14.0.0",
     "@firebase/firestore": "^4.6.1",
     "@hookform/resolvers": "^3.3.4",
-    "@react-native-community/datetimepicker": "7.7.0",
+    "@react-native-community/datetimepicker": "^8.0.0",
     "@react-native-firebase/app": "^19.2.2",
     "@react-native-firebase/firestore": "^19.2.2",
-    "@react-native-firebase/auth": "^19.2.2",
-    "@react-native-vector-icons/ionicons": "^7.3.1-alpha.9",
     "@react-navigation/native": "^6.1.17",
-    "@react-navigation/stack": "^6.3.29",
-    "expo": "~50.0.19",
+    "expo": "~50.0.17",
     "expo-asset": "~9.0.2",
     "expo-font": "~11.10.3",
     "expo-image": "~1.10.6",
@@ -39,28 +36,21 @@
     "react-native": "0.73.6",
     "react-native-element-dropdown": "^2.12.0",
     "react-native-flash-message": "^0.4.2",
-    "react-native-gifted-chat": "^2.4.0",
-    "react-native-mask-input": "^1.2.3",
     "react-native-paper": "^5.12.3",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
     "react-native-simple-toast": "^3.3.1",
     "react-native-toast-message": "^2.2.0",
-    "react-native-vector-icons": "^10.1.0",
     "react-native-web": "~0.19.6",
     "zod": "^3.23.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.45",
-    "@types/react-native-vector-icons": "^6.4.18",
     "jest": "^29.2.1",
     "jest-expo": "~50.0.4",
     "react-test-renderer": "18.2.0",
     "typescript": "^5.1.3"
-  },
-  "overrides": {
-    "expo-constants": "15.4.6"
   },
   "private": true
 }
