+++ src/nutrimetas/shared/LoginSession.tsx
@@ -1,3 +1,5 @@
+"use-strict"
+
 // Dependencies
 // Core React hooks & misc. stuff
 import { ReactNode, Dispatch, createContext, useReducer } from "react";
@@ -10,24 +12,9 @@ export const SessionContext =
 export const SessionDispatchContext = 
     createContext<Dispatch<LoginAction>>(() => {console.log("AAA")});
 
-// User roles on the app
+// Understand the session data and constraints...
 export type UserRole = "professional" | "patient";
-
-// User data on the DB
-export type UserData = {
-    role: UserRole,
-    docId: string,
-    docContents: any,
-}
-
-// User session on the app
-export type LoginSession = {
-    // TODO: Opt for DB schema instead of wildcard key-value type annotations
-    [key: string]: any, 
-    uid: string, // Account's Auth UID
-    docId : string, // Datasheet's DB ID
-    role: UserRole, // Role on the app
-} | undefined;
+export type LoginSession = {[key: string]: any, role? : UserRole} | undefined;
 
 // ... the actions that can be taken on a given session...
 export type LoginAction = {
@@ -42,12 +29,19 @@ function sessionReducer(
     switch (action.type) {
         // Set or replace session if none are active already   
         case 'set': {
-            if (action.newSession === undefined || action.newSession === null)
-            {
-                throw Error(
-                    "Can't set session: New session is nil (Try resetting instead)"
-                );
-            }
+        if (currentSession != undefined && currentSession != null)
+        {
+            throw Error(
+                "Can't set session: Active session already in progress (Try resetting first)"
+            );
+        }
+
+        if (action.newSession == undefined && currentSession != null)
+        {
+            throw Error(
+                "Can't set session: New session is nil (Try resetting instead)"
+            );
+        }
 
             console.log("Setting new session data as", action.newSession);
             return action.newSession;
