+++ src/nutrimetas/tsconfig.json
@@ -6,11 +6,7 @@
       "@/*": [
         "./*"
       ]
-    },
-    "typeRoots": [
-      "node_modules/@types",
-      "src/types"
-    ],
+    }
   },
   "include": [
     "**/*.ts",
