--- src/nutrimetas/types/images.d.ts
@@ -1,5 +1,0 @@
-// Image declarations for ES6's imports 
-declare module "*.png";
-declare module "*.svg";
-declare module "*.jpeg";
-declare module "*.jpg";
